move coreUtils files to utils
