+++ src/h2oMergeFramesInput.js
@@ -1,4 +1,4 @@
-import { uuid } from './coreUtils/uuid';
+import { uuid } from './utils/uuid';
 
 import { flowPreludeFunction } from './flowPreludeFunction';
 const flowPrelude = flowPreludeFunction();
