+++ src/h2oModelInput/h2oModelBuilderForm/h2oModelBuilderForm.js
@@ -4,7 +4,7 @@ import { performValidations } from './performValidations';
 
 import { createControlFromParameter } from '../createControlFromParameter';
 
-import { uuid } from '../../coreUtils/uuid';
+import { uuid } from '../../utils/uuid';
 
 import { flowPreludeFunction } from '../../flowPreludeFunction';
 const flowPrelude = flowPreludeFunction();
