+++ src/h2oPartialDependenceInput.js
@@ -1,5 +1,5 @@
 import { getModelsRequest } from './h2oProxy/getModelsRequest';
-import { uuid } from './coreUtils/uuid';
+import { uuid } from './utils/uuid';
 
 import { flowPreludeFunction } from './flowPreludeFunction';
 const flowPrelude = flowPreludeFunction();
