+++ src/jobOutput/jobOutput.js
@@ -1,6 +1,6 @@
 import { getJobRequest } from '../h2oProxy/getJobRequest';
 import { postCancelJobRequest } from '../h2oProxy/postCancelJobRequest';
-import { formatMilliseconds } from '../coreUtils/formatMilliseconds';
+import { formatMilliseconds } from '../utils/formatMilliseconds';
 
 import { flowPreludeFunction } from '../flowPreludeFunction';
 const flowPrelude = flowPreludeFunction();
