+++ src/utils/getFileBaseName.js
@@ -1,4 +1,4 @@
-import { sanitizeName } from '../coreUtils/sanitizeName';
+import { sanitizeName } from '../utils/sanitizeName';
 
 export function getFileBaseName(filename, extension) {
   return sanitizeName(filename.substr(0, filename.length - extension.length));
