+++ src/utils/sanitizeName.js
